backend_python/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2

# Socket.IO con versioni testate e compatibili
python-socketio[asyncio]==5.11.4
python-engineio==4.9.1
aiofiles==24.1.0